implementing the my account menu and management options from there

Approved-by: Diana M Aschner

+++ backend/src/main/java/com/techelevator/dao/ReviewDao.java
@@ -17,7 +17,7 @@ public interface ReviewDao {
 
     void deleteReview(int reviewId);
 
-    List<Review> getAllReviewsByUserId(int userId);
+    List<Review> getAllReviewsByUserId(long userId);
 
     //List<Beer> topRatedBeers();
